refactor(plugins): move SSL classifiers into PluginKit and drop the test mirror
