+++ Plugins/CassandraDriverPlugin/CassandraConnection.swift
@@ -197,22 +197,8 @@ actor CassandraConnectionActor {
         let keyResult = cass_ssl_set_private_key(ssl, keyString, passphrase)
         if keyResult != CASS_OK {
             cleanup()
-            throw Self.privateKeyLoadError(keyPEM: keyString, hasPassphrase: passphrase != nil, keyPath: keyPath)
+            throw CassandraClientKeyClassifier.privateKeyLoadError(keyPEM: keyString, hasPassphrase: passphrase != nil, keyPath: keyPath)
         }
-    }
-
-    static func isEncryptedPrivateKey(_ pem: String) -> Bool {
-        pem.contains("ENCRYPTED PRIVATE KEY") || (pem.contains("Proc-Type:") && pem.contains("ENCRYPTED"))
-    }
-
-    static func privateKeyLoadError(keyPEM: String, hasPassphrase: Bool, keyPath: String) -> SSLHandshakeError {
-        guard isEncryptedPrivateKey(keyPEM) else {
-            return .clientKeyInvalid(serverMessage: "The client key at \(keyPath) is not a valid private key")
-        }
-        if hasPassphrase {
-            return .clientKeyPassphraseIncorrect(serverMessage: "The passphrase for the client key at \(keyPath) is incorrect")
-        }
-        return .clientKeyPassphraseRequired(serverMessage: "The client key at \(keyPath) is encrypted. Enter its passphrase.")
     }
 
     func close() {
