+++ Plugins/ClickHouseDriverPlugin/ClickHousePlugin.swift
@@ -216,7 +216,7 @@ final class ClickHousePluginDriver: PluginDatabaseDriver, @unchecked Sendable {
             session = nil
             lock.unlock()
             Self.logger.error("Connection test failed: \(error.localizedDescription)")
-            if let sslError = Self.classifySSLError(error) {
+            if let sslError = ClickHouseSSLClassifier.classifySSLError(error) {
                 throw sslError
             }
             throw ClickHouseError.connectionFailed
@@ -708,30 +708,6 @@ final class ClickHousePluginDriver: PluginDatabaseDriver, @unchecked Sendable {
 
     func generateDropIndexSQL(table: String, indexName: String) -> String? {
         "ALTER TABLE \(quoteIdentifier(table)) DROP INDEX \(quoteIdentifier(indexName))"
-    }
-
-    static func classifySSLError(_ error: Error) -> SSLHandshakeError? {
-        let urlError = error as? URLError ?? (error as NSError).underlyingErrors.compactMap { $0 as? URLError }.first
-        if let urlError {
-            switch urlError.code {
-            case .serverCertificateUntrusted, .serverCertificateNotYetValid, .serverCertificateHasUnknownRoot, .serverCertificateHasBadDate:
-                return .untrustedCertificate(serverMessage: urlError.localizedDescription)
-            case .clientCertificateRequired, .clientCertificateRejected:
-                return .clientCertRequired(serverMessage: urlError.localizedDescription)
-            case .secureConnectionFailed:
-                return .cipherMismatch(serverMessage: urlError.localizedDescription)
-            default:
-                break
-            }
-        }
-        let message = error.localizedDescription.lowercased()
-        if message.contains("certificate") && (message.contains("untrusted") || message.contains("verify failed")) {
-            return .untrustedCertificate(serverMessage: error.localizedDescription)
-        }
-        if message.contains("hostname") {
-            return .hostnameMismatch(serverMessage: error.localizedDescription)
-        }
-        return nil
     }
 }
