+++ Plugins/RedisDriverPlugin/RedisPluginConnection.swift
@@ -340,26 +340,6 @@ final class RedisPluginConnection: @unchecked Sendable {
         throw RedisPluginError.hiredisUnavailable
         #endif
     }
-
-    static func classifySSLError(_ message: String) -> SSLHandshakeError? {
-        let lower = message.lowercased()
-        if lower.contains("certificate verify failed") || lower.contains("unable to get local issuer") {
-            return .untrustedCertificate(serverMessage: message)
-        }
-        if lower.contains("hostname") {
-            return .hostnameMismatch(serverMessage: message)
-        }
-        if lower.contains("sslv3") || lower.contains("unsupported protocol") || lower.contains("no shared cipher") {
-            return .cipherMismatch(serverMessage: message)
-        }
-        if lower.contains("ssl handshake failed") || lower.contains("tlsv1") {
-            return .cipherMismatch(serverMessage: message)
-        }
-        if lower.contains("client certificate") {
-            return .clientCertRequired(serverMessage: message)
-        }
-        return nil
-    }
 }
 
 // MARK: - Synchronous Helpers (must be called on the serial queue)
@@ -405,7 +385,7 @@ private extension RedisPluginConnection {
         if result != REDIS_OK {
             redisFreeSSLContext(ssl)
             let errMsg = Self.contextErrorMessage(ctx)
-            if let sslError = Self.classifySSLError(errMsg) {
+            if let sslError = RedisSSLClassifier.classifySSLError(errMsg) {
                 throw sslError
             }
             throw RedisPluginError(code: Int(result), message: "SSL handshake failed: \(errMsg)")
