+++ TableProTests/Plugins/PluginSSLClassifierTests.swift
@@ -1,5 +1,4 @@
 import Foundation
-@testable import TablePro
 import TableProPluginKit
 import Testing
 
@@ -8,7 +7,7 @@ struct LibPQClassifierTests {
     @Test("Classifies the AWS RDS rejection in #1298 as serverRejectedPlaintext")
     func testRDSPattern() {
         let msg = "FATAL: no pg_hba.conf entry for host \"1.2.3.4\", user \"u\", database \"d\", no encryption"
-        guard case .serverRejectedPlaintext = LibPQClassifier.classifySSLError(msg) else {
+        guard case .serverRejectedPlaintext = LibPQSSLClassifier.classifySSLError(msg) else {
             Issue.record("Expected serverRejectedPlaintext")
             return
         }
@@ -17,7 +16,7 @@ struct LibPQClassifierTests {
     @Test("Classifies SSL-required as serverRequiresPlaintext")
     func testSSLRequired() {
         let msg = "FATAL: no pg_hba.conf entry for host \"1.2.3.4\", user \"u\", database \"d\", SSL on"
-        guard case .serverRequiresPlaintext = LibPQClassifier.classifySSLError(msg) else {
+        guard case .serverRequiresPlaintext = LibPQSSLClassifier.classifySSLError(msg) else {
             Issue.record("Expected serverRequiresPlaintext")
             return
         }
@@ -26,7 +25,7 @@ struct LibPQClassifierTests {
     @Test("Classifies server-no-ssl-support as serverRequiresPlaintext")
     func testServerNoSSL() {
         let msg = "server does not support SSL, but SSL was required"
-        guard case .serverRequiresPlaintext = LibPQClassifier.classifySSLError(msg) else {
+        guard case .serverRequiresPlaintext = LibPQSSLClassifier.classifySSLError(msg) else {
             Issue.record("Expected serverRequiresPlaintext")
             return
         }
@@ -35,7 +34,7 @@ struct LibPQClassifierTests {
     @Test("Classifies cert verify failure as untrustedCertificate")
     func testCertVerify() {
         let msg = "SSL error: certificate verify failed"
-        guard case .untrustedCertificate = LibPQClassifier.classifySSLError(msg) else {
+        guard case .untrustedCertificate = LibPQSSLClassifier.classifySSLError(msg) else {
             Issue.record("Expected untrustedCertificate")
             return
         }
@@ -44,7 +43,7 @@ struct LibPQClassifierTests {
     @Test("Classifies hostname mismatch")
     func testHostnameMismatch() {
         let msg = "server certificate for \"foo\" does not match host name \"bar\""
-        guard case .hostnameMismatch = LibPQClassifier.classifySSLError(msg) else {
+        guard case .hostnameMismatch = LibPQSSLClassifier.classifySSLError(msg) else {
             Issue.record("Expected hostnameMismatch")
             return
         }
@@ -52,8 +51,8 @@ struct LibPQClassifierTests {
 
     @Test("Non-SSL error returns nil")
     func testNonSSL() {
-        #expect(LibPQClassifier.classifySSLError("FATAL: password authentication failed") == nil)
-        #expect(LibPQClassifier.classifySSLError("connection refused") == nil)
+        #expect(LibPQSSLClassifier.classifySSLError("FATAL: password authentication failed") == nil)
+        #expect(LibPQSSLClassifier.classifySSLError("connection refused") == nil)
     }
 }
 
@@ -61,7 +60,7 @@ struct LibPQClassifierTests {
 struct MariaDBClassifierTests {
     @Test("CR_SSL_CONNECTION_ERROR with cipher message → cipherMismatch")
     func testSSLConnectionError() {
-        guard case .cipherMismatch = MariaDBClassifier.classifySSLError(code: 2_026, message: "SSL connection error: no shared cipher") else {
+        guard case .cipherMismatch = MariaDBSSLClassifier.classifySSLError(code: 2_026, message: "SSL connection error: no shared cipher") else {
             Issue.record("Expected cipherMismatch")
             return
         }
@@ -69,7 +68,7 @@ struct MariaDBClassifierTests {
 
     @Test("CR_SSL_CONNECTION_ERROR with certificate keyword → untrustedCertificate")
     func testSSLCertError() {
-        guard case .untrustedCertificate = MariaDBClassifier.classifySSLError(code: 2_026, message: "SSL certificate not trusted") else {
+        guard case .untrustedCertificate = MariaDBSSLClassifier.classifySSLError(code: 2_026, message: "SSL certificate not trusted") else {
             Issue.record("Expected untrustedCertificate")
             return
         }
@@ -77,7 +76,8 @@ struct MariaDBClassifierTests {
 
     @Test("require_secure_transport → serverRejectedPlaintext")
     func testRequireSecureTransport() {
-        guard case .serverRejectedPlaintext = MariaDBClassifier.classifySSLError(code: 1_045, message: "Connections using insecure transport are prohibited while --require_secure_transport=ON") else {
+        let message = "Connections using insecure transport are prohibited while --require_secure_transport=ON"
+        guard case .serverRejectedPlaintext = MariaDBSSLClassifier.classifySSLError(code: 1_045, message: message) else {
             Issue.record("Expected serverRejectedPlaintext")
             return
         }
@@ -85,31 +85,12 @@ struct MariaDBClassifierTests {
 
     @Test("Auth error 1045 not retried (returns nil)")
     func testAuthError() {
-        #expect(MariaDBClassifier.classifySSLError(code: 1_045, message: "Access denied for user 'foo'@'bar'") == nil)
+        #expect(MariaDBSSLClassifier.classifySSLError(code: 1_045, message: "Access denied for user 'foo'@'bar'") == nil)
     }
 
     @Test("Network error 2002 not retried")
     func testNetworkError() {
-        #expect(MariaDBClassifier.classifySSLError(code: 2_002, message: "Can't connect to MySQL server") == nil)
-    }
-}
-
-@Suite("FreeTDS SSL Classifier")
-struct FreeTDSClassifierTests {
-    @Test("Server requires encryption → serverRejectedPlaintext")
-    func testServerRequires() {
-        guard case .serverRejectedPlaintext = FreeTDSClassifier.classifySSLError("Server requires encryption") else {
-            Issue.record("Expected serverRejectedPlaintext")
-            return
-        }
-    }
-
-    @Test("OpenSSL handshake → cipherMismatch")
-    func testOpenSSL() {
-        guard case .cipherMismatch = FreeTDSClassifier.classifySSLError("OpenSSL: SSL_connect failed") else {
-            Issue.record("Expected cipherMismatch")
-            return
-        }
+        #expect(MariaDBSSLClassifier.classifySSLError(code: 2_002, message: "Can't connect to MySQL server") == nil)
     }
 }
 
@@ -119,7 +100,7 @@ struct MongoDBClassifierTests {
     func testAtlasInternalErrorHandshake() {
         let message = "No suitable servers found: [TLS handshake failed: internal error (-9838) "
             + "calling hello on 'ac-zmho1ul-shard-00-00.dsllzcf.mongodb.net:27017']"
-        guard case .unknown = MongoDBClassifier.classifySSLError(message) else {
+        guard case .unknown = MongoDBSSLClassifier.classifySSLError(message) else {
             Issue.record("Expected unknown for a generic handshake failure")
             return
         }
@@ -127,7 +108,7 @@ struct MongoDBClassifierTests {
 
     @Test("Genuine cipher/protocol failure → cipherMismatch")
     func testGenuineCipherMismatch() {
-        guard case .cipherMismatch = MongoDBClassifier.classifySSLError("TLS handshake failed: sslv3 alert handshake failure: no shared cipher") else {
+        guard case .cipherMismatch = MongoDBSSLClassifier.classifySSLError("TLS handshake failed: sslv3 alert handshake failure: no shared cipher") else {
             Issue.record("Expected cipherMismatch")
             return
         }
@@ -135,7 +116,7 @@ struct MongoDBClassifierTests {
 
     @Test("Certificate verify failure → untrustedCertificate")
     func testCertificateVerifyFailed() {
-        guard case .untrustedCertificate = MongoDBClassifier.classifySSLError("TLS handshake failed: certificate verify failed") else {
+        guard case .untrustedCertificate = MongoDBSSLClassifier.classifySSLError("TLS handshake failed: certificate verify failed") else {
             Issue.record("Expected untrustedCertificate")
             return
         }
@@ -143,7 +124,7 @@ struct MongoDBClassifierTests {
 
     @Test("Hostname verification failure → hostnameMismatch")
     func testHostnameVerification() {
-        guard case .hostnameMismatch = MongoDBClassifier.classifySSLError("hostname verification failed") else {
+        guard case .hostnameMismatch = MongoDBSSLClassifier.classifySSLError("hostname verification failed") else {
             Issue.record("Expected hostnameMismatch")
             return
         }
@@ -151,7 +132,7 @@ struct MongoDBClassifierTests {
 
     @Test("TLS required → serverRejectedPlaintext")
     func testTLSRequired() {
-        guard case .serverRejectedPlaintext = MongoDBClassifier.classifySSLError("TLS required by Atlas cluster") else {
+        guard case .serverRejectedPlaintext = MongoDBSSLClassifier.classifySSLError("TLS required by Atlas cluster") else {
             Issue.record("Expected serverRejectedPlaintext")
             return
         }
@@ -162,7 +143,7 @@ struct MongoDBClassifierTests {
 struct RedisClassifierTests {
     @Test("No shared cipher → cipherMismatch")
     func testNoSharedCipher() {
-        guard case .cipherMismatch = RedisClassifier.classifySSLError("SSL_connect: no shared cipher") else {
+        guard case .cipherMismatch = RedisSSLClassifier.classifySSLError("SSL_connect: no shared cipher") else {
             Issue.record("Expected cipherMismatch")
             return
         }
@@ -170,7 +151,7 @@ struct RedisClassifierTests {
 
     @Test("Cert verify failed → untrustedCertificate")
     func testCertVerify() {
-        guard case .untrustedCertificate = RedisClassifier.classifySSLError("certificate verify failed (self-signed)") else {
+        guard case .untrustedCertificate = RedisSSLClassifier.classifySSLError("certificate verify failed (self-signed)") else {
             Issue.record("Expected untrustedCertificate")
             return
         }
@@ -181,7 +162,7 @@ struct RedisClassifierTests {
 struct OracleClassifierTests {
     @Test("ORA-29024 → cipherMismatch")
     func testORA29024() {
-        guard case .cipherMismatch = OracleClassifier.classifySSLError("ORA-29024: Certificate validation failure") else {
+        guard case .cipherMismatch = OracleSSLClassifier.classifySSLError("ORA-29024: Certificate validation failure") else {
             Issue.record("Expected cipherMismatch")
             return
         }
@@ -189,12 +170,12 @@ struct OracleClassifierTests {
 
     @Test("Network timeout (ORA-12606) is not classified as SSL")
     func testTimeoutNotSSL() {
-        #expect(OracleClassifier.classifySSLError("ORA-12606: TNS: Application timeout occurred") == nil)
+        #expect(OracleSSLClassifier.classifySSLError("ORA-12606: TNS: Application timeout occurred") == nil)
     }
 
     @Test("ORA-28759 → clientCertRequired")
     func testORA28759() {
-        guard case .clientCertRequired = OracleClassifier.classifySSLError("ORA-28759: failure to open file") else {
+        guard case .clientCertRequired = OracleSSLClassifier.classifySSLError("ORA-28759: failure to open file") else {
             Issue.record("Expected clientCertRequired")
             return
         }
@@ -206,7 +187,7 @@ struct ClickHouseClassifierTests {
     @Test("URLError.secureConnectionFailed → cipherMismatch")
     func testSecureConnectionFailed() {
         let error = URLError(.secureConnectionFailed)
-        guard case .cipherMismatch = ClickHouseClassifier.classifySSLError(error) else {
+        guard case .cipherMismatch = ClickHouseSSLClassifier.classifySSLError(error) else {
             Issue.record("Expected cipherMismatch")
             return
         }
@@ -215,7 +196,7 @@ struct ClickHouseClassifierTests {
     @Test("URLError.serverCertificateUntrusted → untrustedCertificate")
     func testCertUntrusted() {
         let error = URLError(.serverCertificateUntrusted)
-        guard case .untrustedCertificate = ClickHouseClassifier.classifySSLError(error) else {
+        guard case .untrustedCertificate = ClickHouseSSLClassifier.classifySSLError(error) else {
             Issue.record("Expected untrustedCertificate")
             return
         }
@@ -224,7 +205,7 @@ struct ClickHouseClassifierTests {
     @Test("Non-SSL error returns nil")
     func testNonSSL() {
         let error = URLError(.notConnectedToInternet)
-        #expect(ClickHouseClassifier.classifySSLError(error) == nil)
+        #expect(ClickHouseSSLClassifier.classifySSLError(error) == nil)
     }
 }
 
@@ -243,14 +224,14 @@ struct CassandraClassifierTests {
 
     @Test("Detects PKCS#8 and PKCS#1 encrypted keys, not unencrypted ones")
     func testEncryptionDetection() {
-        #expect(CassandraClassifier.isEncryptedPrivateKey(encryptedPkcs8))
-        #expect(CassandraClassifier.isEncryptedPrivateKey(encryptedPkcs1))
-        #expect(!CassandraClassifier.isEncryptedPrivateKey(unencryptedPkcs8))
+        #expect(CassandraClientKeyClassifier.isEncryptedPrivateKey(encryptedPkcs8))
+        #expect(CassandraClientKeyClassifier.isEncryptedPrivateKey(encryptedPkcs1))
+        #expect(!CassandraClientKeyClassifier.isEncryptedPrivateKey(unencryptedPkcs8))
     }
 
     @Test("Encrypted key with no passphrase → clientKeyPassphraseRequired")
     func testEncryptedNoPassphrase() {
-        let error = CassandraClassifier.privateKeyLoadError(
+        let error = CassandraClientKeyClassifier.privateKeyLoadError(
             keyPEM: encryptedPkcs8, hasPassphrase: false, keyPath: "/k.pem")
         guard case .clientKeyPassphraseRequired = error else {
             Issue.record("Expected clientKeyPassphraseRequired")
@@ -260,7 +241,7 @@ struct CassandraClassifierTests {
 
     @Test("Encrypted key with wrong passphrase → clientKeyPassphraseIncorrect")
     func testEncryptedWrongPassphrase() {
-        let error = CassandraClassifier.privateKeyLoadError(
+        let error = CassandraClientKeyClassifier.privateKeyLoadError(
             keyPEM: encryptedPkcs1, hasPassphrase: true, keyPath: "/k.pem")
         guard case .clientKeyPassphraseIncorrect = error else {
             Issue.record("Expected clientKeyPassphraseIncorrect")
@@ -270,9 +251,9 @@ struct CassandraClassifierTests {
 
     @Test("Unencrypted but unreadable key → clientKeyInvalid, never a passphrase error")
     func testUnencryptedInvalid() {
-        let withoutPassphrase = CassandraClassifier.privateKeyLoadError(
+        let withoutPassphrase = CassandraClientKeyClassifier.privateKeyLoadError(
             keyPEM: unencryptedPkcs8, hasPassphrase: false, keyPath: "/k.pem")
-        let withPassphrase = CassandraClassifier.privateKeyLoadError(
+        let withPassphrase = CassandraClientKeyClassifier.privateKeyLoadError(
             keyPEM: unencryptedPkcs8, hasPassphrase: true, keyPath: "/k.pem")
         guard case .clientKeyInvalid = withoutPassphrase else {
             Issue.record("Expected clientKeyInvalid without passphrase")
